Add 'grade local' subcommand for local grading
Easily building the containers is useless, unless you can also run
them locally. The 'grade' subcommand will allow for local execution of
the grader container. The 'local' sub-sub-command instructs the grader
to be run on a submission found on the local file system. In the
future, there will be another sub-sub-command to allow for easy
execution against a true submission downloaded from Coursera.

This commit also cleans up a few tinkercruft things.