Add memory limits on containers.
Enable memory limits for containers started by courseraprogramming in
order to more accurately simulate the production environment. Note:
we ignore memory settings on containers started by the `cat` and `ls`
subcommands, as they should not have any memory or network usage in
the first place.

Further, we allow users to remove memory (and network) restrictions
on containers started from the `inspect` subcommand to allow users to
poke at things in an unrestricted manner. That said, by default we
enable all limits to mimic the real environment as much as possible.