Initial OAuth2 implementation.
This commit is an initial implementation of client-side OAuth2. It
stores information in a local filesystem cache, and automatically
handles retrieving refresh tokens, etc.

After finishing this initial implementation, it appears to make sense
to refactor this into a class, and maybe in the future even into an
independent library.